fix(server): 서버 시작 실패 수정 — 빈 충돌, JWT 로컬 키, bible_books seed, TINYINT 타입
- GetQtUseCaseMock 빈 이름 충돌 해결 (ai/note 도메인 고유 이름 부여)
- application-local.yml에 JWT RS256 테스트 키 추가 (로컬 개발용)
- V7__seed_bible_books.sql 추가 (66권, V8 qt_passages FK 선행 필요)
- V6 ai_validation_logs.layer TINYINT→INT (Entity int 매핑 일치)
- application-local.yml ddl-auto: validate→update (DDL-Entity 불일치 임시 우회)

+++ qtai-server/src/main/resources/db/migration/V6__create_auth_ai_explanation_tables.sql
@@ -134,7 +134,7 @@ CREATE TABLE ai_validation_logs (
     ai_asset_id                 BIGINT       NOT NULL,
     validation_reference_job_id BIGINT,
     checklist_version_id        BIGINT       NOT NULL,
-    layer                       TINYINT      NOT NULL,
+    layer                       INT          NOT NULL,
     result                      VARCHAR(30)  NOT NULL,
     checklist_json              CLOB,
     reviewer_type               VARCHAR(30)  NOT NULL DEFAULT 'AUTO',
